+++ packages/cli/src/build.ts
@@ -1,23 +1,15 @@
 import { CommandModule } from 'yargs';
-import { Docgeni, DocgeniConfig } from '@docgeni/core';
-import { normalizeCommandArgsForAngular } from './angular-args';
+import { Docgeni, DocgeniConfig, readNgBuildOptions } from '@docgeni/core';
 import { getConfiguration } from './configuration';
-import * as fs from 'fs';
-import * as path from 'path';
 import { yargsOptionsGenerate } from './util/yargs-options-generate';
-import { Option } from '@angular/cli/models/interface';
 import { VERSION } from './version';
-const ngBuildOptions: Option[] = JSON.parse(fs.readFileSync(path.resolve(__dirname, './ng-build-options.json')).toString());
+
 export const buildCommand: CommandModule = {
     command: ['build'],
     describe: 'Build documentation site',
     builder: yargs => {
-        yargsOptionsGenerate(yargs, ngBuildOptions)
-            .option('prod', {
-                desc: `Production`,
-                boolean: true,
-                default: false
-            })
+        yargsOptionsGenerate(yargs, readNgBuildOptions());
+        yargs
             .option('skip-site', {
                 desc: `skip build site`,
                 boolean: true,
@@ -37,7 +29,6 @@ export const buildCommand: CommandModule = {
         const docgeni = new Docgeni({
             watch: argv.watch,
             config,
-            cmdArgs: normalizeCommandArgsForAngular(config, ngBuildOptions),
             version: VERSION
         });
         await docgeni.run();