+++ packages/core/src/plugins/plugin.ts
@@ -1,5 +1,9 @@
 import { DocgeniContext } from '../docgeni.interface';
 
-export interface Plugin {
-    apply(docgeni: DocgeniContext): void;
+// export interface Plugin {
+//     apply(docgeni: DocgeniContext): void;
+// }
+
+export abstract class Plugin {
+    abstract apply(docgeni: DocgeniContext): void;
 }