+++ packages/core/src/testing/docgeni-context.ts
@@ -2,12 +2,13 @@ import { toolkit } from '@docgeni/toolkit';
 import { getSystemPath, normalize } from '@angular-devkit/core';
 import { DocgeniPaths } from '../docgeni-paths';
 import { DocgeniContext } from '../docgeni.interface';
-import { DocgeniLibrary } from '../interfaces';
+import { DocgeniConfig, DocgeniLibrary } from '../interfaces';
 import { createTestDocgeniHost } from './docgeni-host';
 import { AsyncSeriesHook, SyncHook } from 'tapable';
 import { DocSourceFile } from '../builders/doc-file';
 import { DocsBuilder } from '../builders';
 import { DocgeniCompilation, LibraryBuilder, LibraryComponent } from '../types';
+import { Docgeni } from '../docgeni';
 
 export const DEFAULT_TEST_ROOT_PATH = normalize(`/D/test`);
 
@@ -49,28 +50,24 @@ export function createTestDocgeniContext(options?: TestDocgeniContextOptions): D
         },
         watch: options.watch,
         paths: paths,
-        hooks: {
-            run: new SyncHook([]),
-            docBuild: new SyncHook<DocSourceFile>(['docSourceFile']),
-            docBuildSucceed: new SyncHook<DocSourceFile>(['docSourceFile']),
-            docsBuild: new SyncHook<DocsBuilder, DocSourceFile[]>(['docsBuilder', 'docs']),
-            docsBuildSucceed: new SyncHook<DocsBuilder, DocSourceFile[]>(['docsBuilder', 'docs']),
-            componentBuild: new SyncHook<LibraryComponent>(['component']),
-            componentBuildSucceed: new SyncHook<LibraryComponent>(['component']),
-            libraryBuild: new SyncHook<LibraryBuilder, LibraryComponent[]>(['libraryBuilder', 'components']),
-            libraryBuildSucceed: new SyncHook<LibraryBuilder, LibraryComponent[]>(['libraryBuilder', 'components']),
-            compilation: new SyncHook<DocgeniCompilation>(['compilation', 'compilationIncrement']),
-            emit: new AsyncSeriesHook<void>([])
-        },
+        hooks: Docgeni.createHooks(),
         logger: toolkit.print,
         librariesBuilder: null,
         docsBuilder: null,
         navsBuilder: null,
         fs: null,
-        enableIvy: null,
+        enableIvy: true,
         compile: () => {
             return Promise.resolve();
         },
         version: '1.0.0'
     };
+}
+
+export function updateContext(context: DocgeniContext, update: Partial<DocgeniContext>) {
+    Object.assign(context, update);
+}
+
+export function updateContextConfig(context: DocgeniContext, update: Partial<DocgeniConfig> | Record<string, any>) {
+    Object.assign(context.config, update);
 }